requirements.txt
tensorflow==2.13.0
flask==2.3.3
opencv-python==4.8.1.78
numpy==1.24.3
pandas==1.5.3
matplotlib==3.7.2
Pillow==10.0.0
scikit-learn==1.3.0
flask-cors==4.0.0
seaborn==0.12.2
# matplotlib.backends.backend_pdf is part of matplotlib
# All other imports are already covered